feat: adds provider to mnemonic screen
