+++ lib/features/landing/widgets/app_drawer_widget.dart
@@ -13,7 +13,7 @@ import 'package:polkadex/features/app_settings_info/screens/terms_conditions_scr
 import 'package:polkadex/features/landing/providers/notification_drawer_provider.dart';
 import 'package:polkadex/features/notifications/screens/notif_deposit_screen.dart';
 import 'package:polkadex/features/notifications/screens/notif_details_screen.dart';
-import 'package:polkadex/features/setup/screens/intro_screen.dart';
+import 'package:polkadex/features/setup/presentation/screens/intro_screen.dart';
 import 'package:polkadex/common/utils/colors.dart';
 import 'package:polkadex/common/utils/enums.dart';
 import 'package:polkadex/common/utils/extensions.dart';
